+++ tests/fixture/keep_platform/namespace/web/input.js
@@ -1,0 +1,10 @@
+import * as env from 'universal-env';
+
+if (env.isWeb) {
+  console.log('This is web');
+} else if (env.isWeex) {
+  console.log('This is weex');
+} else {
+  console.log('others');
+}
+